Refactor weatherobject classes
