+++ index.html
@@ -5,10 +5,12 @@
       rel="stylesheet"
       href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
     />
+    <script src="https://code.jquery.com/jquery-3.4.1.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/chart.js@2.8.0"></script>
     <link rel="stylesheet" href="style.css" />
     <script type="module" src="app.js"></script>
     <script src="node_modules/chart.js/dist/Chart.js"></script>
+
     <title>Charts</title>
   </head>
   <body>
