+++ services/helper-methods.mjs
@@ -1,5 +1,9 @@
 import weatherTableData from '../src/weatherTableData.js';
 import weatherDataObject from '../src/weatherDataObject.js';
+import temperature from '../views/components/temperatureChart.js';
+import humidity from '../views/components/humidityChart.js';
+import pressure from '../views/components/pressureChart.js';
+import all from '../views/components/allCharts.js';
 
 let list = new weatherTableData();
 
@@ -38,6 +42,7 @@ const Utils = {
     return list;
   },
 
+  // Output table with all Data
   outputWeatherObjectsTable: list => {
     let html = '';
     html += `<table>`;
@@ -58,123 +63,14 @@ const Utils = {
     document.getElementById('weather-data').innerHTML = html;
   },
 
+  // Creates a chart based on the available data
   createWeatherChart: response => {
-    let apiData = response;
-    let tempData = [];
-    let humData = [];
-    let pressData = [];
-    let id = [];
+    let data = response;
+    let tempData = data.getTemperatureData();
+    let humData = data.getHumidityData();
+    let pressData = data.getPressureData();
+    let id = data.getDatesData();
 
-    apiData.forEach(res => {
-      tempData.push(res.temperature);
-      humData.push(parseFloat(res.humidity));
-      pressData.push(res.pressure);
-      id.push(new Date(res.date).toLocaleString());
-    });
-
-    let maxHum = humData.reduce((a, b) => {
-      return parseInt(Math.max(a, b));
-    });
-
-    let minHum = humData.reduce((a, b) => {
-      return parseInt(Math.min(a, b));
-    });
-    // let html = `${apiData}`;
-    let all = `
-    var ctx = document.getElementById('all').getContext('2d');
-    var myChart = new Chart(ctx, {
-      type: 'line',
-      data: {
-        labels: id.slice(1, 100),
-        datasets: [
-          {
-            label: 'temperature',
-            borderColor: 'rgba(250, 120, 80, 0.1)',
-            backgroundColor: 'rgba(250, 120, 80, 0.1)',
-            data: tempData.slice(1,100),
-            yAxisID:'A'
-          },{
-            label: 'pressure',
-            borderColor: 'rgba(123, 12, 80, 0.1)',
-            backgroundColor: 'rgba(123, 12, 80, 0.1)',
-            data: pressData.slice(1, 100),
-            yAxisID:'B'
-          },
-          {
-            label: 'humidity',
-            borderColor: 'rgba(50, 59, 80, 0.1)',
-            backgroundColor: 'rgba(50, 59, 80, 0.1)',
-            data: humData.slice(1, 100),
-          }
-        ]
-      },
-      options: {
-        scales: {
-          yAxes: [{
-            id:'A',
-            type:'linear',
-            position:'left',
-            
-          }, {
-            id: 'B',
-            type:'linear',
-            position:'right',
-            ticks:{
-              max:1013,
-              min:1011
-            }
-          }]
-        }
-      }
-    });`;
-
-    let temperature = `
-    var ctx = document.getElementById('temp').getContext('2d');
-    var myChart = new Chart(ctx, {
-      type: 'bar',
-      data: {
-        labels: id,
-        datasets: [
-          {
-            label: 'temperature',
-            data: tempData
-          },
-        ]
-      },
-
-    });`;
-
-    let humidity = `
-    var ctx = document.getElementById('hum').getContext('2d');
-    var myChart = new Chart(ctx, {
-      type: 'bar',
-      data: {
-        labels: id,
-        datasets: [
-          {
-            label: 'humidity',
-            data: humData
-          },
-        ]
-      },
-
-    });`;
-
-    let pressure = `
-    var ctx = document.getElementById('press').getContext('2d');
-    var myChart = new Chart(ctx, {
-      type: 'line',
-      data: {
-        labels: id,
-        datasets: [
-          {
-            label: 'humidity',
-            data: pressData
-          },
-        ]
-      },
-
-    });`;
     // document.getElementById('weather-data').innerHTML = html;
     document.getElementById('all').innerHTML = eval(all);
     document.getElementById('temp').innerHTML = eval(temperature);
