+++ src/weatherDataObject.js
@@ -26,6 +26,4 @@ class WeatherDataObject {
   }
 }
 
-module.exports = WeatherDataObject;
-
-// export default WeatherDataObject;
+export default WeatherDataObject;
